+++ sourcecode/Runtime/PWLambda.cpp
@@ -6,15 +6,15 @@ namespace Nom
 {
 	namespace Runtime
 	{
-		llvm::Value* PWLambda::ReadField(NomBuilder& builder, int32_t fieldIndex)
+		llvm::Value* PWLambda::ReadField(NomBuilder& builder, size_t fieldIndex)
 		{
 			return MakeInvariantLoad(builder, REFTYPE, builder->CreateGEP(LambdaHeader::GetLLVMType(), wrapped, { MakeInt32(0), MakeInt32(LambdaHeaderFields::Fields), MakeInt32(fieldIndex) }));
 		}
-		llvm::Value* PWLambda::WriteField(NomBuilder& builder, int32_t fieldIndex, llvm::Value* value)
+		llvm::Value* PWLambda::WriteField(NomBuilder& builder, size_t fieldIndex, llvm::Value* value)
 		{
 			return MakeInvariantStore(builder, value, builder->CreateGEP(LambdaHeader::GetLLVMType(), wrapped, { MakeInt32(0), MakeInt32(LambdaHeaderFields::Fields), MakeInt32(fieldIndex) }));
 		}
-		llvm::Value* PWLambdaPrecise::ReadLambdaField(NomBuilder& builder, int32_t fieldIndex)
+		llvm::Value* PWLambdaPrecise::ReadLambdaField(NomBuilder& builder, size_t fieldIndex)
 		{
 			return MakeInvariantLoad(builder, REFTYPE, builder->CreateGEP(LambdaHeader::GetLLVMType(), wrapped, { MakeInt32(0), MakeInt32(LambdaHeaderFields::Fields), MakeInt32(fieldIndex) }));
 		}
