+++ sourcecode/Runtime/instructions/DebugInstruction.h
@@ -14,7 +14,7 @@ namespace Nom
 		public:
 			DebugInstruction(const ConstantID message);
 			virtual ~DebugInstruction() override;
-			virtual void Compile(NomBuilder &builder, CompileEnv* env, int lineno) override;
+			virtual void Compile(NomBuilder &builder, CompileEnv* env, size_t lineno) override;
 
 			// Inherited via NomInstruction
 			virtual void Print(bool resolve = false) override;
