+++ sourcecode/Runtime/instructions/LoadBoolConstantInstruction.h
@@ -13,7 +13,7 @@ namespace Nom
 			LoadBoolConstantInstruction(RegIndex reg, bool value);
 
 			// Inherited via NomValueInstruction
-			virtual void Compile(NomBuilder& builder, CompileEnv* env, int lineno) override;
+			virtual void Compile(NomBuilder& builder, CompileEnv* env, size_t lineno) override;
 			virtual void Print(bool resolve = false) override;
 
 			virtual void FillConstantDependencies(NOM_CONSTANT_DEPENCENCY_CONTAINER& result) override;
